refactor: split out the linter into its own file
