+++ packages/cspell/src/fileHelper.ts
@@ -1,0 +1,90 @@
+import * as cspell from 'cspell-lib';
+import * as fsp from 'fs-extra';
+import getStdin from 'get-stdin';
+import { GlobOptions, globP } from './util/glob';
+import * as path from 'path';
+import { CSpellUserSettings } from 'cspell-lib';
+
+const UTF8: BufferEncoding = 'utf8';
+const STDIN = 'stdin';
+
+export interface ConfigInfo {
+    source: string;
+    config: CSpellUserSettings;
+}
+
+export interface FileConfigInfo {
+    configInfo: ConfigInfo;
+    filename: string;
+    text: string;
+    languageIds: string[];
+}
+
+export async function readConfig(configFile: string | undefined, root: string | undefined): Promise<ConfigInfo> {
+    if (configFile) {
+        const config = (await cspell.loadConfig(configFile)) || {};
+        return { source: configFile, config };
+    }
+    const config = await cspell.searchForConfig(root);
+    return { source: config?.__importRef?.filename || 'None found', config: config || {} };
+}
+
+export interface FileInfo {
+    filename: string;
+    text: string;
+}
+
+export function readFileInfo(filename: string, encoding: string = UTF8): Promise<FileInfo> {
+    const pText = filename === STDIN ? getStdin() : fsp.readFile(filename, encoding);
+    return pText.then(
+        (text) => ({ text, filename }),
+        (error) => {
+            return error.code === 'EISDIR'
+                ? Promise.resolve({ text: '', filename })
+                : Promise.reject({
+                      ...error,
+                      message: `Error reading file: "${filename}"`,
+                  });
+        }
+    );
+}
+
+export function readFile(filename: string, encoding: string = UTF8): Promise<string> {
+    return readFileInfo(filename, encoding).then((info) => info.text);
+}
+
+/**
+ * Looks for matching glob patterns or stdin
+ * @param globPatterns patterns or stdin
+ */
+export async function findFiles(globPatterns: string[], options: GlobOptions): Promise<string[]> {
+    const globPats = globPatterns.filter((filename) => filename !== STDIN);
+    const stdin = globPats.length < globPatterns.length ? [STDIN] : [];
+    const globResults = globPats.length ? await globP(globPats, options) : [];
+    const cwd = options.cwd || process.cwd();
+    return stdin.concat(globResults.map((filename) => path.resolve(cwd, filename)));
+}
+
+export function calcFinalConfigInfo(
+    configInfo: ConfigInfo,
+    settingsFromCommandLine: CSpellUserSettings,
+    filename: string,
+    text: string
+): FileConfigInfo {
+    const ext = path.extname(filename);
+    const fileSettings = cspell.calcOverrideSettings(configInfo.config, path.resolve(filename));
+    const settings = cspell.mergeSettings(
+        cspell.getDefaultSettings(),
+        cspell.getGlobalSettings(),
+        fileSettings,
+        settingsFromCommandLine
+    );
+    const languageIds = settings.languageId ? [settings.languageId] : cspell.getLanguagesForExt(ext);
+    const config = cspell.constructSettingsForText(settings, text, languageIds);
+    return {
+        configInfo: { ...configInfo, config },
+        filename,
+        text,
+        languageIds,
+    };
+}
